feat: Instructors page now get data from API
