+++ AssignmentManager/Client/Pages/Index.razor
@@ -1,22 +1,22 @@
 ﻿@page "/"
 @using MudBlazor
 
-<h1>Hello, world!</h1>
-
-Welcome to your new app.
-
-<SurveyPrompt Title="How is Blazor working for you?" />
-
-<MudText Typo="Typo.h6">Next Random number is @Numb</MudText>
-<MudButton Variant="Variant.Filled" Color="Color.Primary" OnClick="OnButtonsClick">@ButtonText</MudButton>
-<MudButton Variant="Variant.Outlined" Color="Color.Primary" OnClick="OnButtonsClick">@ButtonText</MudButton>
-<MudButton Variant="Variant.Text" Color="Color.Primary" OnClick="OnButtonsClick">@ButtonText</MudButton>
-@code {
-    public int Numb { get; set; } = 0;
-    public string ButtonText { get; set; } = "Click Me";
-    private Random _rand = new Random();
-    void OnButtonsClick()
-    {
-        Numb = _rand.Next();
-    }
-}
+<h1>Добро пожаловать в Систему Упрвления Отчётами!</h1>
+<h2>Возможно, вам будут интересны, следующие страницы</h2>
+<ul class="nav flex-column">
+    <li class="nav-item px-3">
+        <NavLink class="nav-link" href="" Match="NavLinkMatch.All">
+            <span class="oi oi-home" aria-hidden="true"></span> Home
+        </NavLink>
+    </li>
+    <li class="nav-item px-3">
+        <NavLink class="nav-link" href="specialities">
+            <span class="oi oi-list-rich" aria-hidden="true"></span> Specialities
+        </NavLink>
+    </li>
+    <li class="nav-item px-3">
+        <NavLink class="nav-link" href="Instructors">
+            <span class="oi oi-list-rich" aria-hidden="true"></span> Instructors
+        </NavLink>
+    </li>
+</ul>
