+++ AssignmentManager/Client/Shared/NavMenu.razor
@@ -13,16 +13,6 @@
           </NavLink>
       </li>
       <li class="nav-item px-3">
-          <NavLink class="nav-link" href="counter">
-              <span class="oi oi-plus" aria-hidden="true"></span> Counter
-          </NavLink>
-      </li>
-      <li class="nav-item px-3">
-          <NavLink class="nav-link" href="fetchdata">
-              <span class="oi oi-list-rich" aria-hidden="true"></span> Fetch data
-          </NavLink>
-      </li>
-      <li class="nav-item px-3">
           <NavLink class="nav-link" href="specialities">
               <span class="oi oi-list-rich" aria-hidden="true"></span> Specialities
           </NavLink>
