+++ AssignmentManager/Server/AssignmentManager.Server.csproj
@@ -34,6 +34,7 @@
 
   <ItemGroup>
     <Folder Include="Data\Migrations" />
+    <Folder Include="Resources\" />
   </ItemGroup>
 
 </Project>
