+++ AssignmentManager/Server/Controllers/InstructorsController.cs
@@ -3,11 +3,10 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using AssignmentManager.Server.Extensions;
 using AssignmentManager.Server.Models;
-using AssignmentManager.Server.Resources;
+using AssignmentManager.Shared;
 using AssignmentManager.Server.Services;
 using AutoMapper;
 using Microsoft.AspNetCore.Mvc;
-using AssignmentManager.Server.Extensions;
 
 namespace AssignmentManager.Server.Controllers
 {
