+++ AssignmentManager/Server/Controllers/SubjectsController.cs
@@ -5,11 +5,10 @@ using AssignmentManager.Server.Extensions;
 using Microsoft.AspNetCore.Mvc;
 using AssignmentManager.Server.Models;
 using AssignmentManager.Server.Persistence.Contexts;
-using AssignmentManager.Server.Resources;
+using AssignmentManager.Shared;
 using AssignmentManager.Server.Services;
 using AutoMapper;
 using IdentityModel.Client;
-using AssignmentManager.Server.Extensions;
 
 namespace AssignmentManager.Server.Controllers
 {
