+++ AssignmentManager/Shared/InstructorResource.cs
@@ -1,18 +1,16 @@
-﻿using System.ComponentModel.DataAnnotations;
+﻿using System.Collections.Generic;
 
-namespace AssignmentManager.Server.Resources
+namespace AssignmentManager.Shared
 {
-    public class SaveInstructorResource
+    public class InstructorResource
     {
-        [Required]
         public int IsuId { get; set; }
-        [Required]
         public string LastName { get; set; }
-        [Required]
         public string FirstName { get; set; }
         public string PatronymicName { get; set; }
-        [Required]
         public string Email { get; set; }
         public string Phone { get; set; }
+
+        public IList<SubjectResource> Subjects { get; set; }
     }
 }
