+++ AssignmentManager/Shared/SaveInstructorResource.cs
@@ -1,16 +1,18 @@
-﻿using System.Collections.Generic;
+﻿using System.ComponentModel.DataAnnotations;
 
-namespace AssignmentManager.Server.Resources
+namespace AssignmentManager.Shared
 {
-    public class InstructorResource
+    public class SaveInstructorResource
     {
+        [Required]
         public int IsuId { get; set; }
+        [Required]
         public string LastName { get; set; }
+        [Required]
         public string FirstName { get; set; }
         public string PatronymicName { get; set; }
+        [Required]
         public string Email { get; set; }
         public string Phone { get; set; }
-
-        public IList<SubjectResource> Subjects { get; set; }
     }
 }
